requirements.txt
aliyun-python-sdk-core==2.13.32
aliyun-python-sdk-kms==2.14.0
asn1crypto==0.24.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cloud-init==19.1.7
command-not-found==0.3
configobj==5.0.6
crcmod==1.7
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
distro-info==0.18ubuntu0.18.04.1
gensim==3.6.0
idna==2.8
imageio==2.9.0
importlib-metadata==3.4.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
MarkupSafe==1.1.1
matplotlib==3.2.1
netifaces==0.10.4
networkx==2.5
nltk==3.4.5
numpy==1.18.2
oauthlib==3.1.0
oss2==2.14.0
pandas==1.0.3
Pillow==8.1.2
pip==21.0.1
protobuf==3.15.6
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pygobject==3.26.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
requests==2.22.0
scikit-image==0.17.2
scikit-learn==0.20.1
scipy==1.1.0
screen-resolution-extra==0.0.0
seaborn==0.7.1
SecretStorage==2.3.1
setuptools==52.0.0
six==1.11.0
sklearn==0.0
smart-open==4.2.0
ssh-import-id==5.7
tensorboardX==2.1
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
typing-extensions==3.7.4.3
ufw==0.36
urllib3==1.25.6
wheel==0.30.0
xkit==0.0.0
zipp==3.4.0
